+++ src/eth_provider/database/mod.rs
@@ -1,3 +1,5 @@
+pub mod types;
+
 use futures::TryStreamExt;
 use mongodb::{
     bson::Document,
