+++ src/eth_provider/database/types/header.rs
@@ -4,6 +4,6 @@ use serde::Deserialize;
 /// A header as stored in the database
 #[derive(Debug, Deserialize)]
 pub struct StoredHeader {
-    #[serde(deserialize_with = "crate::eth_provider::types::serde::deserialize_intermediate")]
+    #[serde(deserialize_with = "crate::eth_provider::database::types::serde::deserialize_intermediate")]
     pub header: Header,
 }
