+++ src/eth_provider/database/types/receipt.rs
@@ -4,7 +4,7 @@ use serde::Deserialize;
 /// A transaction receipt as stored in the database
 #[derive(Debug, Deserialize)]
 pub struct StoredTransactionReceipt {
-    #[serde(deserialize_with = "crate::eth_provider::types::serde::deserialize_intermediate")]
+    #[serde(deserialize_with = "crate::eth_provider::database::types::serde::deserialize_intermediate")]
     pub receipt: TransactionReceipt,
 }
