+++ src/eth_provider/mod.rs
@@ -1,6 +1,5 @@
 pub mod database;
 pub mod error;
-pub mod kakarot;
 pub mod provider;
-pub mod types;
+pub mod starknet;
 pub mod utils;
